+++ zhmccli/__init__.py
@@ -35,3 +35,7 @@ from ._cmd_storagegroup import *  # noqa: F401
 from ._cmd_storagevolume import *  # noqa: F401
 from ._cmd_vstorageresource import *  # noqa: F401
 from ._cmd_capacitygroup import *  # noqa: F401
+from ._cmd_user import *       # noqa: F401
+from ._cmd_user_role import *  # noqa: F401
+from ._cmd_password_rule import *  # noqa: F401
+from ._cmd_character_rule import *  # noqa: F401